🎨 Decouple content styles from blog slug component
