+++ static/global.css
@@ -49,7 +49,7 @@ body {
     @apply font-serif;
 }
 
-.main > * + *, [class$="-body"] > * + * {
+.main > * + *, .content > * + * {
     margin-top: var(--s2);
 }
